[libsignon-qt] Hard-code fallback p2p socket directory. Contributes to JB#42126
Previously, the code assumed that the client and server were running
in the same context, so that the p2p dbus socket dir returned by:
QStandardPaths::writableLocation(QStandardPaths::RuntimeLocation)
would be the same for each.  This is generally not the case (e.g.,
signond will be run with privileges by invoker, etc), and this would
cause the client library to fall back to session dbus connection.

This commit ensures that the appropriate fallback directory is
specified within the client library so that p2p dbus is used.

Contributes to JB#42126